fix(cli): measure durations with a monotonic clock
Wall-clock timestamps (now_ms) are vulnerable to NTP corrections and DST
shifts, which can produce negative or skewed durations. Iteration and
measurement timing now bracket with a monotonic clock instead.

Also consolidates the duplicated budget-summary/trailer logic in
loop_cmds.py, adds a typed JsonRenderer protocol, narrows swallowed
exceptions in shared.py from bare Exception to GymratError | OSError,
and centralizes minute-to-ms conversion into session/budget.py.

+++ src/gymrat/report/json_doc.py
@@ -51,7 +51,7 @@ class BudgetSummary:
     Computed at report time so the renderer stays clock-free.
     """
 
-    cap_minutes: int
+    cap_minutes: float
     remaining_seconds: int
